Move run_cargo to Builder
